+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Transport 5.X.X
 
+1. Added log::Recorder::Filename that returns the name of the log file.
+   * [Pull request 327](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/327)
+
 1. Added a logging tutorial
    * [Pull request 311](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/311)